+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # embetty
 
+## Unreleased
+
+- Links are now opened in a new tab.
+
 ## v1.0.0-beta.7, 2018-01-29
 
 - Added missing `skip_cleanup: true` to Travis CI config.